+++ src/components/index.js
@@ -1,0 +1,3 @@
+import WelcomePlaceholder from "./welcome-placeholder/WelcomePlaceholder";
+
+export { WelcomePlaceholder };
