--- src/pages/LandingPage.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-import WelcomeScreen from "../components/components/WelcomeScreen/WelcomeScreen";
-
-function LandingPage() {
-	return (
-		<div>
-			<WelcomeScreen />
-		</div>
-	);
-}
-
-export default LandingPage;
